- reuse addSecrets for modified secrets - check for equality of annotations
